+++ test/built-ins/TypedArray/prototype/subarray/BigInt/speciesctor-get-species-custom-ctor-returns-another-instance.js
@@ -37,7 +37,7 @@ features: [BigInt, Symbol.species, TypedArray]
 
 testWithBigIntTypedArrayConstructors(function(TA) {
   var sample = new TA([40n]);
-  var other = new Int8Array([1, 0, 1]);
+  var other = new BigInt64Array([1n, 0n, 1n]);
   var result;
 
   sample.constructor = {};
@@ -48,5 +48,5 @@ testWithBigIntTypedArrayConstructors(function(TA) {
   result = sample.subarray(0, 0);
 
   assert.sameValue(result, other, "returned another typedarray");
-  assert(compareArray(result, [1, 0, 1]), "the returned object is preserved");
+  assert(compareArray(result, [1n, 0n, 1n]), "the returned object is preserved");
 });
